+++ test/components/subhead_component_test.rb
@@ -13,7 +13,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_renders_heading
     render_inline(Primer::SubheadComponent.new) do |component|
-      component.slot(:heading, tag: :h2) { "Hello World" }
+      component.heading(tag: :h2) { "Hello World" }
     end
 
     assert_selector(".Subhead h2.Subhead-heading", text: "Hello World")
@@ -21,7 +21,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_render_dangerous_heading
     render_inline(Primer::SubheadComponent.new) do |component|
-      component.slot(:heading, danger: true) { "Hello World" }
+      component.heading(danger: true) { "Hello World" }
     end
 
     assert_selector(".Subhead .Subhead-heading--danger", text: "Hello World")
@@ -29,7 +29,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_render_without_border
     render_inline(Primer::SubheadComponent.new(hide_border: true)) do |component|
-      component.slot(:heading) { "Hello World" }
+      component.heading { "Hello World" }
     end
 
     assert_selector(".Subhead.border-bottom-0.mb-0", text: "Hello World")
@@ -37,7 +37,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_bottom_margin_can_be_overridden_when_border_is_hidden
     render_inline(Primer::SubheadComponent.new(hide_border: true, mb: 1)) do |component|
-      component.slot(:heading) { "Hello World" }
+      component.heading { "Hello World" }
     end
 
     assert_selector(".Subhead.border-bottom-0.mb-1", text: "Hello World")
@@ -45,8 +45,8 @@ class SubheadComponentTest < Minitest::Test
 
   def test_renders_actions
     render_inline(Primer::SubheadComponent.new(heading: "Hello world")) do |component|
-      component.slot(:heading) { "Hello World" }
-      component.slot(:actions) { "My Actions" }
+      component.heading { "Hello World" }
+      component.actions { "My Actions" }
     end
 
     assert_selector(".Subhead .Subhead-actions", text: "My Actions")
@@ -54,7 +54,7 @@ class SubheadComponentTest < Minitest::Test
 
   def test_handles_spacious
     render_inline(Primer::SubheadComponent.new(spacious: true)) do |component|
-      component.slot(:heading) { "Hello World" }
+      component.heading { "Hello World" }
     end
 
     assert_selector(".Subhead.Subhead--spacious .Subhead-heading", text: "Hello World")
@@ -62,8 +62,8 @@ class SubheadComponentTest < Minitest::Test
 
   def test_renders_a_description
     render_inline(Primer::SubheadComponent.new(heading: "Hello world")) do |component|
-      component.slot(:heading) { "Hello World" }
-      component.slot(:description) { "My Description" }
+      component.heading { "Hello World" }
+      component.description { "My Description" }
     end
 
     assert_selector(".Subhead .Subhead-description", text: "My Description")